accel/amdxdna: Make hmm_invalidate ops common for AIE2 and AIE4
Move the hmm_invalidate into common aie.c code, both AIE2 and AIE4
can share the same operation.

Co-developed-by: Hayden Laccabue <Hayden.Laccabue@amd.com>
